cf_load_firebase/requirements.txt
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0")
certifi==2022.9.24; python_version >= "3.7" and python_version < "3.11"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0"
google-api-core==2.10.2; python_version >= "3.7" and python_version < "3.11"
google-auth==2.13.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0"
google-cloud-appengine-logging==1.1.6; python_version >= "3.7"
google-cloud-audit-log==0.2.4; python_version >= "3.7"
google-cloud-bigquery-storage==2.16.2; python_version >= "3.7" and python_version < "3.11"
google-cloud-bigquery==3.3.5; python_version >= "3.7" and python_version < "3.11"
google-cloud-core==2.3.2; python_version >= "3.7" and python_version < "3.11"
google-cloud-firestore==2.7.2; python_version >= "3.7"
google-cloud-logging==3.2.5; python_version >= "3.7"
google-crc32c==1.5.0; python_version >= "3.7" and python_version < "3.11"
google-resumable-media==2.4.0; python_version >= "3.7" and python_version < "3.11"
googleapis-common-protos==1.56.4; python_version >= "3.7" and python_version < "3.11"
grpc-google-iam-v1==0.12.4; python_version >= "3.7"
grpcio-status==1.50.0; python_version >= "3.7" and python_version < "3.11"
grpcio==1.50.0; python_version >= "3.7" and python_version < "3.11"
idna==3.4; python_version >= "3.7" and python_version < "3.11"
numpy==1.23.4; python_version >= "3.8" and python_version < "3.11"
packaging==21.3; python_version >= "3.7" and python_version < "3.11"
proto-plus==1.22.1; python_version >= "3.7" and python_version < "3.11"
protobuf==4.21.9; python_version >= "3.7" and python_version < "3.11"
pyarrow==9.0.0; python_version >= "3.7" and python_version < "3.11"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0") or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0")
pyparsing==3.0.9; python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.8"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.3.0"
requests==2.28.1; python_version >= "3.7" and python_version < "3.11"
rsa==4.9; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.6.0"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version < "3.11" and python_version >= "3.7"
